Cros commit

+++ app/Http/Controllers/NotificationController.php
@@ -14,7 +14,7 @@ class NotificationController extends Controller
     		$user = JWTAuth::parseToken()->toUser();
     		if($user instanceof User) {
     			$notification = Notification::find($id);
-    			if($notification instanceof Notification) {
+    			if($notification instanceof Notification && $notification->fellowship_id == $user->fellowship_id) {
     				return response()->json(['notification' => $notification], 200);
     			} else {
     				return response()->json(['error' => 'notification not found'], 404);
@@ -30,11 +30,11 @@ class NotificationController extends Controller
     	try {
     		$user = JWTAuth::parseToken()->toUser();
     		if($user instanceof User) {
-    			$count = Notification::count();
+    			$count = Notification::where([['is_seen', '=', 0], ['fellowship_id', '=', $user->fellowship_id]])->count();
     			if($count == 0) {
     				return response()->json(['notification' => 0], 404);
     			} 
-    			$notifications = Notification::paginate(10);
+    			$notifications = Notification::where('fellowship_id', '=', $user->fellowship_id)->paginate(10);
     			return response()->json(['notifications' => $notifications, 'count' => $count], 200);
     		} else {
     			return response()->json(['error' => 'token expired'], 401);
@@ -48,7 +48,7 @@ class NotificationController extends Controller
     		$user = JWTAuth::parseToken()->toUser();
     		if($user instanceof User) {
     			$notification = Notification::find($id);
-    			if($notification instanceof Notification) {
+    			if($notification instanceof Notification && $notification->fellowship_id == $user->fellowship_id) {
     				if($notification->delete()) {
     					return response()->json(['message' => 'notification removed successfully'], 200);
     				} else {
@@ -61,6 +61,22 @@ class NotificationController extends Controller
     			return response()->json(['error' => 'token expired'], 401);
     		}
     	} catch(Exception $ex) {
+            return response()->json(['message' => 'somthing went wrong', 'error' => $ex->getMessage()], $ex->getStatusCode());
+        }
+    }
+    public function seenNotification(Request $request) {
+        try {
+            $user = JWTAuth::parseToken()->toUser();
+            if($user instanceof User) {
+                $notification = Notification::find($request->input("id"));
+                if($notification instanceof Notification) {
+                    $notification->is_seen = true;
+                    $notification->update();
+                }
+            } else {
+                return response()->json(['error' => 'token expired'], 401);
+            }
+        } catch(Exception $ex) {
             return response()->json(['message' => 'somthing went wrong', 'error' => $ex->getMessage()], $ex->getStatusCode());
         }
     }
